Rename GamePresenter.h to game-source-code/GamePresenter.h
